Street crime/requirements.txt
Flask==2.3.2
Flask-Login==0.7.0
Flask-SocketIO==6.5.3
python-socketio==5.11.0
eventlet==0.33.3
Flask-SQLAlchemy==3.0.3
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.9
redis==4.5.5
python-dotenv==1.0.0
gunicorn==20.1.0
